feat(sentinel): <- adds http pathway to see bpm calculation

+++ v3_bridges/sentinel/src/lib/utils.rs
@@ -6,5 +6,5 @@ use std::{
 use crate::SentinelError;
 
 pub fn get_utc_timestamp() -> Result<u64, SentinelError> {
-    Ok(SystemTime::now().duration_since(UNIX_EPOCH)?.as_secs() as u64)
+    Ok(SystemTime::now().duration_since(UNIX_EPOCH)?.as_secs())
 }
